# insomnia/openapi.yaml
openapi: 3.0.3
info:
  title: Neon Life Drawing API
  description: |
    API for managing life drawing classes, models, venues, and calendar bookings.
    
    ## Authentication
    Most endpoints require a JWT Bearer token. Obtain one via the `/api/auth/login` or `/api/auth/register` endpoints.
    
    ## Base URLs
    - Production: `https://neon-api-production.fliptopbox.workers.dev`
    - Local: `http://localhost:8787`
  version: 1.0.0
  contact:
    name: Neon API Support
servers:
  - url: https://neon-api-production.fliptopbox.workers.dev
    description: Production server
  - url: http://localhost:8787
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login or register
  schemas:
...
    Venue:
      type: object
      properties:
        id:
          type: integer
        user_id:
          type: integer
        name:
          type: string
        week_day:
          type: integer
          description: "0: Sunday, 1: Monday, ..., 6: Saturday"
        frequency:
          type: string
          enum: [weekly, biweekly, monthly]
        instagram:
          type: string
        website:
          type: string
        address:
          type: string
        timezone:
          type: string
        start_time:
          type: string
          description: Time in HH:MM format
        duration:
          type: integer
          description: Duration in minutes
        postcode:
          type: string
        area:
          type: string
        price_inperson:
          type: integer
          description: Price in pence
        price_online:
          type: integer
          description: Price in pence
        active:
          type: integer
        tags:
          type: array
          items:
            type: string
        created_on:
          type: string
          format: date-time
        modified_on:
          type: string
          format: date-time
        organizer_name:
          type: string
...
paths:
...
  /api/venues:
    get:
      tags:
        - Venues
      summary: List all active venues
      responses:
        '200':
          description: List of venues
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Venue'
    post:
      tags:
        - Venues
      summary: Create venue
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
                - week_day
                - start_time
                - duration
              properties:
                name:
                  type: string
                week_day:
                  type: integer
                frequency:
                  type: string
                  default: weekly
                instagram:
                  type: string
                website:
                  type: string
                address:
                  type: string
                timezone:
                  type: string
                start_time:
                  type: string
                duration:
                  type: integer
                postcode:
                  type: string
                area:
                  type: string
                price_inperson:
                  type: integer
                price_online:
                  type: integer
                tags:
                  type: array
                  items:
                    type: string
            example:
              name: Life Drawing at The Studio
              week_day: 3
              frequency: weekly
              start_time: "19:00"
              duration: 180
              address: 123 Art Street
              postcode: E1 1AA
              area: East London
              price_inperson: 1500
              price_online: 1000
              tags: ["clothed", "portrait"]
      responses:
        '201':
          description: Venue created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '401':
          description: Unauthorized
  /api/venues/{id}:
    get:
      tags:
        - Venues
      summary: Get venue by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Venue details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '404':
          description: Venue not found
    put:
      tags:
        - Venues
      summary: Update venue
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                week_day:
                  type: integer
                frequency:
                  type: string
                instagram:
                  type: string
                website:
                  type: string
                address:
                  type: string
                timezone:
                  type: string
                start_time:
                  type: string
                duration:
                  type: integer
                postcode:
                  type: string
                area:
                  type: string
                price_inperson:
                  type: integer
                price_online:
                  type: integer
                tags:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Venue updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Venue'
        '401':
          description: Unauthorized
        '404':
          description: Venue not found
    delete:
      tags:
        - Venues
      summary: Delete venue (Admin only, soft delete)
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Venue deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
        '401':
          description: Unauthorized
        '403':
          description: Forbidden - Admin access required

  # Calendar endpoints
...